get rid of the ugly hack
Previously, it took a long time to generate voice.
As a countermeasure,
while the user was selecting the save destination,
the generation process was proceeding behind the scenes.
However, even if you canceled the dialog,
the generation process proceeded and there was no way to interrupt it.
This behavior was not intuitive to the user.

Now it's faster than before.
This ugly hack is no longer needed.